Consolidate logging module into components
