+++ nautilus_trader/adapters/betfair/client.py
@@ -58,7 +58,7 @@ from betfair_parser.spec.navigation import Menu
 from betfair_parser.spec.navigation import Navigation
 
 import nautilus_trader
-from nautilus_trader.common.logging import Logger
+from nautilus_trader.common.component import Logger
 from nautilus_trader.core.nautilus_pyo3 import HttpClient
 from nautilus_trader.core.nautilus_pyo3 import HttpMethod
 from nautilus_trader.core.nautilus_pyo3 import HttpResponse
