+++ nautilus_trader/adapters/bybit/websocket/client.py
@@ -19,8 +19,8 @@ import json
 from collections.abc import Callable
 
 from nautilus_trader.common.clock import LiveClock
+from nautilus_trader.common.component import Logger
 from nautilus_trader.common.enums import LogColor
-from nautilus_trader.common.logging import Logger
 from nautilus_trader.core.nautilus_pyo3 import WebSocketClient
 from nautilus_trader.core.nautilus_pyo3 import WebSocketConfig
