+++ nautilus_trader/adapters/interactive_brokers/client/common.py
@@ -29,8 +29,8 @@ from ibapi.execution import Execution
 from nautilus_trader.adapters.interactive_brokers.common import IBContract
 from nautilus_trader.cache.cache import Cache
 from nautilus_trader.common.clock import LiveClock
+from nautilus_trader.common.component import Logger
 from nautilus_trader.common.component import MessageBus
-from nautilus_trader.common.logging import Logger
 from nautilus_trader.model.data import BarType
 from nautilus_trader.model.identifiers import InstrumentId
