+++ nautilus_trader/adapters/interactive_brokers/historic/client.py
@@ -32,10 +32,10 @@ from nautilus_trader.adapters.interactive_brokers.providers import InteractiveBr
 # fmt: on
 from nautilus_trader.cache.cache import Cache
 from nautilus_trader.common.clock import LiveClock
+from nautilus_trader.common.component import Logger
 from nautilus_trader.common.component import MessageBus
-from nautilus_trader.common.logging import Logger
-from nautilus_trader.common.logging import init_logging
-from nautilus_trader.common.logging import log_level_from_str
+from nautilus_trader.common.component import init_logging
+from nautilus_trader.common.component import log_level_from_str
 from nautilus_trader.core.datetime import dt_to_unix_nanos
 from nautilus_trader.core.datetime import unix_nanos_to_dt
 from nautilus_trader.model.data import Bar
