+++ nautilus_trader/backtest/engine.pyx
@@ -47,12 +47,12 @@ from nautilus_trader.common.clock cimport LiveClock
 from nautilus_trader.common.clock cimport TestClock
 from nautilus_trader.common.clock cimport TimeEvent
 from nautilus_trader.common.clock cimport TimeEventHandler
-from nautilus_trader.common.logging cimport Logger
-from nautilus_trader.common.logging cimport log_level_from_str
-from nautilus_trader.common.logging cimport log_sysinfo
-from nautilus_trader.common.logging cimport set_logging_clock_realtime_mode
-from nautilus_trader.common.logging cimport set_logging_clock_static_mode
-from nautilus_trader.common.logging cimport set_logging_clock_static_time
+from nautilus_trader.common.component cimport Logger
+from nautilus_trader.common.component cimport log_level_from_str
+from nautilus_trader.common.component cimport log_sysinfo
+from nautilus_trader.common.component cimport set_logging_clock_realtime_mode
+from nautilus_trader.common.component cimport set_logging_clock_static_mode
+from nautilus_trader.common.component cimport set_logging_clock_static_time
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.data cimport Data
 from nautilus_trader.core.datetime cimport maybe_dt_to_unix_nanos
