+++ nautilus_trader/common/actor.pyx
@@ -41,12 +41,12 @@ from libc.stdint cimport uint64_t
 from nautilus_trader.cache.base cimport CacheFacade
 from nautilus_trader.common.clock cimport Clock
 from nautilus_trader.common.clock cimport LiveClock
+from nautilus_trader.common.component cimport CMD
+from nautilus_trader.common.component cimport REQ
+from nautilus_trader.common.component cimport SENT
 from nautilus_trader.common.component cimport Component
+from nautilus_trader.common.component cimport Logger
 from nautilus_trader.common.component cimport MessageBus
-from nautilus_trader.common.logging cimport CMD
-from nautilus_trader.common.logging cimport REQ
-from nautilus_trader.common.logging cimport SENT
-from nautilus_trader.common.logging cimport Logger
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.data cimport Data
 from nautilus_trader.core.message cimport Event
