+++ nautilus_trader/system/kernel.py
@@ -33,14 +33,14 @@ from nautilus_trader.common.actor import Actor
 from nautilus_trader.common.clock import Clock
 from nautilus_trader.common.clock import LiveClock
 from nautilus_trader.common.clock import TestClock
+from nautilus_trader.common.component import Logger
 from nautilus_trader.common.component import MessageBus
+from nautilus_trader.common.component import init_logging
+from nautilus_trader.common.component import init_tracing
+from nautilus_trader.common.component import log_header
 from nautilus_trader.common.enums import LogColor
 from nautilus_trader.common.enums import LogLevel
 from nautilus_trader.common.enums import log_level_from_str
-from nautilus_trader.common.logging import Logger
-from nautilus_trader.common.logging import init_logging
-from nautilus_trader.common.logging import init_tracing
-from nautilus_trader.common.logging import log_header
 from nautilus_trader.config import ActorFactory
 from nautilus_trader.config import DataEngineConfig
 from nautilus_trader.config import ExecEngineConfig
